requirements.txt
aiohttp==3.4.4
docker==3.5.1
PyYAML==3.13